+++ kele.el
@@ -101,8 +101,11 @@ pods."
   :type 'integer
   :group 'kele)
 
-(defvar kele-after-context-switch-hook nil
-  "Normal hook run after switching to a new context.")
+(defvar kele-context-after-switch-functions nil
+  "Functions to run after switching to a new context.
+
+Each function is expected to take a single string argument
+representing the name of the new context.")
 
 (defvar kele--discovery-last-refresh-time nil
   "Timestamp of last successful poll of the discovery cache.")
@@ -136,7 +139,7 @@ NAMESPACE GVK NAME PROCESS).")
   "Which major mode to use for YAML highlighting.
 
 Set to nil to disable YAML highlighting."
-  :type '(choice nil symbol)
+  :type '(choice (const nil) symbol))
   :group 'kele)
 
 (defcustom kele-max-column-width 40
@@ -151,7 +154,7 @@ Set to nil to disable YAML highlighting."
               "Found multiple group-versions associated with the given resource")
 
 (defface kele-disabled-face
-  '((t (:inherit 'font-lock-comment-face)))
+  '((t (:inherit font-lock-comment-face)))
   "Face used for disabled or not-applicable values."
   :group 'kele)
 
@@ -436,12 +439,12 @@ If CONTEXT is nil, use the current context."
   "Look up the namespaced-ness of GROUP-VERSION TYPE in CACHE.
 
 If CONTEXT is not provided, the current context is used."
-  (if-let ((namespaced-p
-            (->> (kele--get-resource-lists-for-context cache (or context (kele-current-context-name)))
-                 (-first (lambda (resource-list) (equal (alist-get 'groupVersion resource-list) group-version)))
-                 (alist-get 'resources)
-                 (-first (lambda (resource) (equal (alist-get 'name resource) type)))
-                 (alist-get 'namespaced))))
+  (if-let* ((namespaced-p
+             (->> (kele--get-resource-lists-for-context cache (or context (kele-current-context-name)))
+                  (-first (lambda (resource-list) (equal (alist-get 'groupVersion resource-list) group-version)))
+                  (alist-get 'resources)
+                  (-first (lambda (resource) (equal (alist-get 'name resource) type)))
+                  (alist-get 'namespaced))))
       (not (eq :false namespaced-p))
     (signal 'kele-cache-lookup-error `(,context ,group-version ,type))))
 
@@ -692,8 +695,7 @@ returns nil."
 (cl-defmethod proxy-active-p ((manager kele--proxy-manager)
                               context)
   "Return non-nil if a proxy serve is active for CONTEXT in MANAGER."
-  (when-let (res (assoc context (oref manager records)))
-    (cdr res)))
+  (cdr (assoc context (oref manager records))))
 
 (defvar kele--global-proxy-manager (kele--proxy-manager))
 
@@ -769,8 +771,8 @@ to complete.  Returned value may not be up to date."
 
 (defun kele--context-cluster-name (context-name)
   "Get the name of the cluster of the context named CONTEXT-NAME."
-  (if-let ((context (-first (lambda (elem) (string= (alist-get 'name elem) context-name))
-                            (alist-get 'contexts (oref kele--global-kubeconfig-cache contents)))))
+  (if-let* ((context (-first (lambda (elem) (string= (alist-get 'name elem) context-name))
+                             (alist-get 'contexts (oref kele--global-kubeconfig-cache contents)))))
       (alist-get 'cluster (alist-get 'context context))
     (error "Could not find context of name %s" context-name)))
 
@@ -891,7 +893,8 @@ node `(elisp)Programmed Completion'."
   (interactive (list (completing-read "Context: " #'kele--contexts-complete)))
   (kele--with-progress (format "Switching to use context `%s'..." context)
     (kele-kubectl-do "config" "use-context" context)
-    (run-hooks 'kele-after-context-switch-hook)))
+    (run-hooks 'kele-after-context-switch-hook)
+    (run-hook-with-args 'kele-context-after-switch-functions context)))
 
 ;; TODO(#176): Update `kele--namespace-cache'
 (transient-define-suffix kele-context-rename (old-name new-name)
@@ -968,7 +971,7 @@ If value is nil, the namespaces need to be fetched directly.")
 If the namespaces are cached, return the cached value.
 
 If CACHE is non-nil, cache the fetched namespaces."
-  (if-let ((cached-namespaces (alist-get (intern context) kele--namespaces-cache)))
+  (if-let* ((cached-namespaces (alist-get (intern context) kele--namespaces-cache)))
       cached-namespaces
     (let* ((gvk (kele--gvk-create :version "v1" :kind "namespaces"))
            (namespaces (kele--fetch-resource-names gvk :context context)))
@@ -1289,10 +1292,8 @@ context and namespace in its name."
                      :resource (kele--resource-container-resource object)
                      :namespace (kele--resource-container-namespace object)))
         (put 'kele--current-resource-buffer-context 'permanent-local t))
-
       (when kele-yaml-highlighting-mode
         (funcall kele-yaml-highlighting-mode))
-
       (kele-get-mode 1))
     (select-window (display-buffer buf))))
 
@@ -1302,12 +1303,12 @@ context and namespace in its name."
         (curr alist))
     (dolist (key keys)
       (setq prev curr)
-      (setq curr (if-let ((list-p (listp curr))
-                          (res (assq key curr)))
+      (setq curr (if-let* ((list-p (listp curr))
+                           (res (assq key curr)))
                      (cdr res)
                    nil)))
     (when curr
-      (assq-delete-all (car (last keys)) prev))
+      (ignore (assq-delete-all (car (last keys)) prev)))
     alist))
 
 (defvar kele--context-keymap nil
@@ -1438,7 +1439,7 @@ Assumes that the current Transient prefix's :scope is an alist w/ `context' key.
   ;; FIXME: Make this resilient to the prefix's scope not having a context
   ;; value.  If not present (or the scope is not an alist or the scope is not
   ;; defined), default to current context.
-  (if-let ((context (alist-get 'context (oref transient--prefix scope))))
+  (if-let* ((context (alist-get 'context (oref transient--prefix scope))))
       (kele--namespaces-complete
        :context context
        :prompt prompt
@@ -1484,10 +1485,10 @@ Also resets any specified peer arguments on the same prefix that
   match any element of `:resettees' on OBJ."
   (cl-call-next-method obj val)
   (dolist (arg (oref obj resettees))
-    (when-let ((obj (cl-find-if (lambda (obj)
-                                  (and (slot-boundp obj 'argument)
-                                       (equal (oref obj argument) arg)))
-                                transient--suffixes)))
+    (when-let* ((obj (cl-find-if (lambda (obj)
+                                   (and (slot-boundp obj 'argument)
+                                        (equal (oref obj argument) arg)))
+                                 transient--suffixes)))
       (transient-init-value obj))))
 
 (defclass kele--transient-infix (kele--transient-infix-resetter
@@ -1611,10 +1612,9 @@ Defaults to the currently active context as set in
 
 First checks the current Transient command's arguments if set.
 Otherwise, returns the current context name from kubeconfig."
-  (if-let* ((cmd transient-current-command)
-            (args (transient-args cmd))
-            (value (transient-arg-value "--context=" args)))
-      value
+  (if transient-current-command
+      (transient-arg-value "--context=" (transient-args
+                                         transient-current-command))
     (kele-current-context-name)))
 
 (cl-defun kele--get-namespace-arg (&key (permit-nil nil) use-default group-version kind (prompt "Namespace: "))
@@ -1633,9 +1633,10 @@ In order of priority, this function attempts the following:
   namespaced, return nil;
 
 - Otherwise, ask the user to select a namespace using PROMPT."
-  (let ((transient-arg-maybe (->> transient-current-command
-                                  (transient-args)
-                                  (transient-arg-value "--namespace="))))
+  (let ((transient-arg-maybe (when transient-current-command
+                               (->> transient-current-command
+                                    (transient-args)
+                                    (transient-arg-value "--namespace=")))))
     (cond
      ((or transient-arg-maybe permit-nil) transient-arg-maybe)
      ((or (not (and group-version kind)) use-default)